bots/telegram/pyproject.toml
[project]
name = "hyacine-gallery-bot-telegram"
version = "0.1.0"
description = "Telegram bot worker for Hyacine Gallery"
requires-python = ">=3.14"
dependencies = [
    "python-telegram-bot[job-queue]>=20.5",
    "httpx>=0.28",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "Pillow>=11.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9",
    "mypy>=1.14",
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "respx>=0.22",
]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"
